+++ src/main/resources/templates/layout.html
@@ -13,7 +13,7 @@
     <nav>
         <ul>
             <li><a th:href="@{/}">시작</a></li>
-            <li><a th:href="@{/profile}">홈</a></li>
+            <li><a th:href="@{/user/profile}">홈</a></li>
             <li><a th:href="@{/question/write}">질문 등록</a></li>
             <li><a th:href="@{/question/list}">질문 목록</a></li>
